requirements.txt
openai==0.27.8
python-dotenv
tweepy
redis